# Followers/followers-be/docs/swagger.yaml
basePath: /
host: localhost:8082
info:
  contact: {}
  description: API for following users
  title: Followers API
  version: "1.0"
paths:
  /api/followers/{username}/follow:
    post:
      description: Authenticated user follows another user
      parameters:
      - description: Username to follow
        in: path
        name: username
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Follow user
      tags:
      - followers
  /api/followers/{username}/unfollow:
    delete:
      description: Authenticated user unfollows another user
      parameters:
      - description: Username to unfollow
        in: path
        name: username
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Unfollow user
      tags:
      - followers
  /api/followers/check/{username}:
    get:
      description: Checks if authenticated user follows another user
      parameters:
      - description: Username to check
        in: path
        name: username
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Check following
      tags:
      - followers
  /api/followers/following:
    get:
      description: Get users that authenticated user follows
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get following users
      tags:
      - followers
  /api/followers/recommendations:
    get:
      description: Get follower recommendations
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get recommendations
      tags:
      - followers
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"